✨ Feat: 보고서 제목에 볼드 텍스트 포맷팅 기능 추가 및 Breadcrumb 컴포넌트에서 아이콘 주석 처리
