+++ src/utils/formatBoldText.tsx
@@ -1,0 +1,11 @@
+export const formatBoldText = (text: string) => {
+  const parts = text.split(/(\*\*[^*]+\*\*)/g);
+
+  return parts.map((part, index) => {
+    if (part.startsWith("**") && part.endsWith("**")) {
+      const boldText = part.slice(2, -2);
+      return <b key={index}>{boldText}</b>;
+    }
+    return part;
+  });
+};
